[WFCORE-XXX] Override resolved attribute value from env var
When resolving the value for an attribute, check if there is an env var
that corresponds to the "global" of the attribute (i.e. resource path adress + name of the attribute). If so, override the resolved value with
content of the env var.

The env var to check is based on the resource address (minus the leading
`/`) and the name of the attribute (all in upper case).

For example, the attribute `inet-address` on the `/interface=management`
resource can be overridden using the env var
`INTERFACE_MANAGEMENT_INET_ADDRESS`.

By default:

```
$ ./bin/standalone.sh
...
15:25:54,655 INFO  [org.jboss.as] (Controller Boot Thread) WFLYSRV0060: Http management interface listening on http://127.0.0.1:9990/management
...
```

Override with an env var:

```
$ INTERFACE_MANAGEMENT_INET_ADDRESS=0.0.0.0 ./bin/standalone.sh
...
15:30:17,507 INFO  [org.jboss.as] (Controller Boot Thread) WFLYSRV0060: Http management interface listening on http://0.0.0.0:9990/management
```

Signed-off-by: Jeff Mesnil <jmesnil@redhat.com>